Move debug stuff under cranelift-wasm/src/debug
