+++ Cargo.toml
@@ -20,7 +20,6 @@ path = "src/clif-util.rs"
 [dependencies]
 cfg-if = "0.1"
 cranelift-codegen = { path = "cranelift-codegen", version = "0.30.0" }
-cranelift-debug = { path = "cranelift-debug", version = "0.30.0", optional = true }
 cranelift-entity = { path = "cranelift-entity", version = "0.30.0" }
 cranelift-reader = { path = "cranelift-reader", version = "0.30.0" }
 cranelift-frontend = { path = "cranelift-frontend", version = "0.30.0" }
@@ -44,9 +43,8 @@ pretty_env_logger = "0.3.0"
 file-per-thread-logger = "0.1.2"
 
 [features]
-default = ["disas", "wasm", "dwarf"]
+default = ["disas", "wasm"]
 disas = ["capstone"]
-dwarf = ["cranelift-debug", "wasm"]
 wasm = ["wabt", "cranelift-wasm"]
 
 # We want debug symbols on release binaries by default since it allows profiling
