--- cranelift-debug/README.md
@@ -1,4 +1,0 @@
-This is the `cranelift-debug` crate, which provides functionality to
-read, transform, and write DWARF section.
-
-[`cranelift-debug`]: https://crates.io/crates/cranelift-debug
